Fix a bug with `select()` introduced by the anti-Zalgo commit
Make sealed thenables Zalgo-releasing, as otherwise there is a window of uncertainty between settling of a thenable and notification of `select()` logic. If it happens that select succeeds with some operation, and this moment overlaps with uncertainty window of some other completed operation, the logic of select breaks, as it expects that any operation is either pending (and may be cancelled in this case) or completed.