Re-work merge to not be based on #maybeCan{Take,Send}Sync()
This was a bad idea, since #maybeCan{Take,Send}Sync() have unintended properties and side effects (e.g. triggering other send/take waiters and, in turn, the ability to be triggered by them), which makes correctly implementing merge based on these functions really hard.

Probably we'll need to revisit this later as a part of work on remote channels, as they can't provide all guarantees needed for the implementation introduced in this commit to work.